Added a Haskell data type that wraps up IplImage.
The HIplImage data type includes the features from the IplImage struct
that OpenCV makes use of. This simplifies the job of inspecting an
IplImage obtained from OpenCV for further processing in Haskell.

Added the PixelUtils module for working with OpenCV's image packing
style. The primary issue is that the image data buffer in an IplImage
will sometimes contain room for four color channels even if only three
are used. This is good for considering an image on a row-by-row basis,
but complicates the view of image data as a single vector.