Added the packPixels function.
This replaces the old dropAlpha as it is faster and more general.
The idea is to trim OpenCV images that have a stride that is
larger than strictly necessary. Such images have their valid
pixel data copied and placed into a Vector that can be sent
on the network faster than the padded version.