Normalized all unsafe rewrites to use the pipeline function.
pipeline is a function defined in HIplUtils that passes a
duplicated image to a supplied function. The critical piece
is that pipeline itself has rewrite rules that collapse
compositions of pipelined functions so that the duplication
only occurs at the head of the composite function.

The idea is that the processing functions are typically implemented
by allocating a fresh image into which their results are stored. In
the case where one of the source parameters could serve as a
destination, the processing function creates a compatible image
before writing into it. These functions are referentially transparent
as the allocation is not observable with respect to the semantics of
the program. The unsafe variants do not themselves allocate a
destination image, so, to preserve referential transparency, we must
ensure that the overwritten source image is not observable. This is
accomplished by first wrapping all unsafe operations in a pipeline
call (which creates a duplicate image for the unsafe operation to
clobber), but then fusing composed pipelines to reduce the number
of allocations.