Converted unsafe operations to live in IO. pipeline does the cast.
By leaving the unsafe operations in the IO monad, the strictness of
the pipeline function is more clear. Previously, a strictness
annotation was required in order to make explicit the dependency
of a processing operation on the image data it is supplied to deal
with the effects of calling unsafePerformIO on that image data.

With pipeline now using standard monadic sequencing, all the unsafe
operations are wrapped with the same unsafePerformIO, so no extra
annotation is needed to ensure the correct strictness properties.