Improved the in-place mutation deforestation mechanisms.
Operations that can be performed in-place give rise to a pair of
functions: safe and unsafe. They are also equipped with a pair of
rewrite rules. The first rule rewrites a safe application into a
@pipeline (unsafe args...)@ call. The @pipeline@ function has
associated rewrite rules that collapse compositions of pipelined
functions. If a pipeline is not collapsed, then the second rule
associated with the original function can kick in. This rule
transforms @pipeline (unsafe args...)@ to @safe args...@.

Unfortunately, the phase control on these rules interacts with the
inliner in a fairly complex manner. To avoid un-deforested pipeline
calls, user code should be annotated {-# INLINE [2] foo #-}. This
annotation has the effect of letting both the pipeline rewrite fire in
an early phase, and then the unpipe rewrite to fire at a later phase
if necessary. Without this annotation, the worst case scenario is that
a function that could work with an uninitialized freshly allocated
buffer instead gets a buffer whose contents have been initialized via
a copy from an existing image.